fix: dpToPx 返回 Float 修复 Paint.textSize 类型不匹配

+++ app/src/main/java/com/ck66/dusou/overlay/RegionSelectOverlay.kt
@@ -326,7 +326,7 @@ private class RegionSelectView(context: Context) : View(context) {
         return this * resources.displayMetrics.density
     }
 
-    private fun Int.dpToPx(): Int {
-        return (this * resources.displayMetrics.density).toInt()
+    private fun Int.dpToPx(): Float {
+        return this * resources.displayMetrics.density
     }
 }
